pyproject.toml
[tool.poetry]
name = "telegram-search"
version = "0.1.0"
description = ""
authors = ["DimaSimakov <dimarodeo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.33.1"
python-dotenv = "^1.0.0"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
